+++ tiptap/src/components/tiptap-command-groups/index.ts
@@ -1,5 +1,6 @@
 export * from "./lists-command-group";
 export * from "./text-align-command-group";
+export * from "./text-decoration-command-group";
 
 export * from "./make-folio-tiptap-nodes-command-group";
 export * from "./make-layouts-command-group";