+++ tiptap/src/components/tiptap-ui/folio-tiptap-node-button/index.tsx
@@ -1,2 +1,1 @@
 export * from "./folio-tiptap-node-button";
-export * from "./folio-tiptap-node-button-for-single-image";